requirements.txt
django==2.1
psycopg2==2.7.5
gunicorn==19.9.0
django-filter==2.0.0
djangorestframework==3.8.2
dj-static==0.0.6
static==1.1.1
static3==0.7.0
virtualenv==16.0.0
whitenoise==4.0
pystache==0.5.4
pytz==2018.5